pyproject.toml
[project]
name = "ai-scientist-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Shunsuke KITADA", email = "shunsuke.kitada.0831@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "grandalf>=0.8",
    "langchain[openai]>=0.3.25",
    "langchain-community>=0.3.24",
    "langgraph>=0.4.7",
    "loguru>=0.7.3",
    "pillow>=11.2.1",
    "pydantic-settings>=2.9.1",
    "semanticscholar>=0.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.0.0",
    "pytest>=6.0.0",
    "ruff>=0.1.5",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ai_scientist_v2"]
